move objectlist which lists s3 from prune to objects change release naming convention to make scheduler loading simpler
